Move duplication to separate modal
